PHPCS lint fixes
