+++ tests/Api/FederationTest.php
@@ -1,4 +1,5 @@
 <?php
+
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestFactoryInterface;
 use Psr\Http\Message\StreamFactoryInterface;
@@ -265,4 +266,3 @@ test('Federation::deleteFederation', function () {
 
     expect($result)->toBe($responseData);
 });
-
