+++ tests/Api/StopwordsTest.php
@@ -1,4 +1,5 @@
 <?php
+
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestFactoryInterface;
 use Psr\Http\Message\StreamFactoryInterface;
